рефакторинг, базовый фрагмент PictureOfTheDay с методом увеличения размера картинки
